requirements.txt
appdirs==1.4.3
certifi==2019.6.16
chardet==3.0.4
cssselect==1.1.0
Django==2.2.5
django-background-tasks==1.2.0
django-compat==1.0.15
django-extensions==2.2.1
django-nine==0.2.2
djangorestframework==3.10.3
facebook-scraper==0.1.4
fake-useragent==0.1.11
future==0.17.1
idna==2.8
jsonfield==2.0.2
lxml==4.4.1
oauthlib==3.1.0
parse==1.12.1
praw==6.3.1
prawcore==1.0.1
pyee==6.0.0
pyppeteer==0.0.25
pyquery==1.4.0
PySocks==1.7.0
python-dateutil==2.8.0
python-twitter==3.5
pytz==2019.2
requests==2.22.0
requests-html==0.10.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
tqdm==4.36.1
tweepy==3.8.0
update-checker==0.16
urllib3==1.25.3
w3lib==1.21.0
websocket-client==0.56.0
websockets==8.0.2
yattag==1.12.2
